# docs/swagger.yaml
definitions:
  ClientError:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/ErrCode'
        example: SERVER_NOT_RESPONDING
    type: object
  ErrCode:
    enum:
    - SERVER_INTERNAL_ERROR
    - SERVER_NOT_RESPONDING
    - USER_INVALID_DATA
    - USER_ALREADY_EXISTS
    - USER_NOT_FOUND
    - AUTH_INVALID_TOKEN
    - AUTH_EXPIRED_TOKEN
    - AUTH_TOKEN_NOT_PROVIDED
    - AUTH_INVALID_EMAIL_OR_PASSWORD
    type: string
    x-enum-varnames:
    - ErrServerInternalError
    - ErrServerNotResponding
    - ErrUserInvalidData
    - ErrUserAlreadyExists
    - ErrUserNotFound
    - ErrAuthInvalidToken
    - ErrAuthExpiredToken
    - ErrAuthTokenNotProvided
    - ErrAuthInvalidEmailOrPassword
  HttpOutput:
    properties:
      data:
        additionalProperties: {}
        type: object
    type: object
  LoginWithEmailPasswordInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: StrongPassword1234
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  SignUpWithEmailPasswordInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: StrongPassword1234
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  UpdateMeInput:
    properties:
      name:
        example: John Doe
        type: string
      password:
        example: securepassword
        maxLength: 72
        minLength: 8
        type: string
      phone:
        example: "09123456789"
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: AliMoradi0Business@gmail.com
    name: Ali Moradi
  description: Toppira APIs Documents.
  title: Toppira APIs
  version: "1.0"
paths:
  /auth/login-with-user-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/LoginWithEmailPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      summary: login with email and password
      tags:
      - Authentication
  /auth/sign-up-with-user-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign Up Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SignUpWithEmailPasswordInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/HttpOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      summary: sign up with email and password
      tags:
      - Authentication
  /user/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: get my(user) data
      tags:
      - User
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update Me Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateMeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: update my(user) data
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"